Restore original ZBS global value for each test
This is a requirement in hopes of having just some backwards compatibility with existing tests.
The demo REST API tests have also been updated to no longer use mocks, so we have real tests that create and fetch contacts.
